Updated auth pages. Removed less code becouse a dry principle was violated. Updated file structure. Added second input for repeat password. Updated styles for SignForm.
