+++ frontend/src/components/signPages/form.components.scss
@@ -1,9 +1,10 @@
-@import "../../../config.scss";
+@import "../../config.scss";
 
 .form {
   background-color: #ffffff;
   max-width: 454px;
-  max-height: 600px;
+  height: fit-content;
+  max-height: 700px;
   min-height: 600px;
   position: absolute;
   top: 0;
