+++ frontend/src/components/signPages/formError.components.scss
@@ -1,4 +1,4 @@
-@import "../../../config.scss";
+@import "../../config.scss";
 
 .form__error {
   font-weight: normal;
