+++ frontend/src/components/signPages/passwordInput.components.tsx
@@ -1,0 +1,85 @@
+import React, { useState } from "react";
+
+export const PasswordInput = ({ page, placeholder, id }: { page: string, placeholder: string, id: string }) => {
+  const [hidePassword, setHidePassword] = useState(true);
+
+  return (
+    <div id="passwordContainer" className="form__container form__password_margin ">
+      <label htmlFor="passwordInput">
+        <svg
+          width="18"
+          height="20"
+          viewBox="0 0 18 20"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <path
+            d="M14.8333 7.01999V5.83333C14.8333 4.28623 14.2188 2.8025 13.1248 1.70854C12.0308 0.614576 10.5471 -5.72205e-06 9 -5.72205e-06C7.4529 -5.72205e-06 5.96917 0.614576 4.87521 1.70854C3.78125 2.8025 3.16667 4.28623 3.16667 5.83333V7.01999C2.42447 7.34391 1.79274 7.87709 1.34875 8.55433C0.904752 9.23157 0.667729 10.0235 0.666664 10.8333V15.8333C0.667987 16.938 1.1074 17.997 1.88851 18.7781C2.66963 19.5593 3.72867 19.9987 4.83333 20H13.1667C14.2713 19.9987 15.3304 19.5593 16.1115 18.7781C16.8926 17.997 17.332 16.938 17.3333 15.8333V10.8333C17.3323 10.0235 17.0952 9.23157 16.6513 8.55433C16.2073 7.87709 15.5755 7.34391 14.8333 7.01999ZM4.83333 5.83333C4.83333 4.72826 5.27232 3.66845 6.05372 2.88705C6.83512 2.10565 7.89493 1.66666 9 1.66666C10.1051 1.66666 11.1649 2.10565 11.9463 2.88705C12.7277 3.66845 13.1667 4.72826 13.1667 5.83333V6.66666H4.83333V5.83333ZM15.6667 15.8333C15.6667 16.4964 15.4033 17.1322 14.9344 17.6011C14.4656 18.0699 13.8297 18.3333 13.1667 18.3333H4.83333C4.17029 18.3333 3.53441 18.0699 3.06556 17.6011C2.59672 17.1322 2.33333 16.4964 2.33333 15.8333V10.8333C2.33333 10.1703 2.59672 9.5344 3.06556 9.06556C3.53441 8.59672 4.17029 8.33333 4.83333 8.33333H13.1667C13.8297 8.33333 14.4656 8.59672 14.9344 9.06556C15.4033 9.5344 15.6667 10.1703 15.6667 10.8333V15.8333Z"
+            fill="#ACACAC"
+          />
+          <path
+            d="M9 11.6667C8.77898 11.6667 8.56702 11.7545 8.41074 11.9107C8.25446 12.067 8.16666 12.279 8.16666 12.5V14.1667C8.16666 14.3877 8.25446 14.5996 8.41074 14.7559C8.56702 14.9122 8.77898 15 9 15C9.22101 15 9.43297 14.9122 9.58925 14.7559C9.74553 14.5996 9.83333 14.3877 9.83333 14.1667V12.5C9.83333 12.279 9.74553 12.067 9.58925 11.9107C9.43297 11.7545 9.22101 11.6667 9 11.6667Z"
+            fill="#ACACAC"
+          />
+        </svg>
+      </label>
+      <input
+        id={id}
+        className="form__input form__password form__password_error"
+        type={
+          page === 'logIn' ?
+            hidePassword ? "password" : "text"
+            : 'text'
+        }
+        placeholder={placeholder}
+      />
+      <div
+        className="passwordToogler passwordToogler_error"
+        onClick={() => {
+          if (hidePassword) {
+            setHidePassword(false);
+          } else {
+            setHidePassword(true);
+          }
+        }}
+      >
+        {page === 'logIn' ?
+          hidePassword ? (
+            <svg
+              width="20"
+              height="20"
+              viewBox="0 0 20 20"
+              fill="none"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                d="M19.3925 7.84928C18.6402 6.61739 17.6912 5.51699 16.5833 4.59178L18.9166 2.25845C19.0684 2.10128 19.1524 1.89078 19.1505 1.67228C19.1486 1.45379 19.061 1.24478 18.9065 1.09027C18.752 0.935762 18.543 0.848122 18.3245 0.846223C18.106 0.844324 17.8955 0.92832 17.7383 1.08012L15.2008 3.62095C13.6278 2.68663 11.8295 2.19967 9.99998 2.21262C4.84081 2.21262 1.90081 5.74428 0.607482 7.84928C0.207922 8.49553 -0.00372314 9.24031 -0.00372314 10.0001C-0.00372314 10.7599 0.207922 11.5047 0.607482 12.1509C1.35978 13.3828 2.30872 14.4832 3.41665 15.4084L1.08331 17.7418C1.00372 17.8186 0.940238 17.9106 0.896564 18.0123C0.85289 18.1139 0.829901 18.2233 0.82894 18.3339C0.827978 18.4446 0.849063 18.5543 0.890964 18.6567C0.932864 18.7591 0.994742 18.8522 1.07299 18.9304C1.15123 19.0087 1.24427 19.0706 1.34669 19.1125C1.4491 19.1544 1.55883 19.1754 1.66948 19.1745C1.78013 19.1735 1.88948 19.1505 1.99115 19.1069C2.09282 19.0632 2.18478 18.9997 2.26165 18.9201L4.80498 16.3768C6.37604 17.3109 8.17221 17.7987 9.99998 17.7876C15.1591 17.7876 18.0991 14.2559 19.3925 12.1509C19.792 11.5047 20.0037 10.7599 20.0037 10.0001C20.0037 9.24031 19.792 8.49553 19.3925 7.84928ZM2.02748 11.2784C1.79009 10.8943 1.66436 10.4517 1.66436 10.0001C1.66436 9.54855 1.79009 9.10591 2.02748 8.72178C3.13915 6.91678 5.65165 3.87928 9.99998 3.87928C11.3835 3.87153 12.7477 4.20498 13.9716 4.85012L12.2941 6.52762C11.4941 5.99645 10.5349 5.75845 9.57934 5.85399C8.62378 5.94953 7.7307 6.37273 7.05165 7.05178C6.3726 7.73084 5.94939 8.62391 5.85385 9.57947C5.75831 10.535 5.99632 11.4942 6.52748 12.2943L4.60248 14.2193C3.58173 13.394 2.71068 12.3992 2.02748 11.2784ZM12.5 10.0001C12.5 10.6632 12.2366 11.299 11.7677 11.7679C11.2989 12.2367 10.663 12.5001 9.99998 12.5001C9.62874 12.4987 9.26265 12.4132 8.92915 12.2501L12.25 8.92928C12.4131 9.26278 12.4985 9.62887 12.5 10.0001ZM7.49998 10.0001C7.49998 9.33707 7.76337 8.70119 8.23221 8.23235C8.70105 7.76351 9.33694 7.50011 9.99998 7.50011C10.3712 7.50155 10.7373 7.58702 11.0708 7.75011L7.74998 11.0709C7.58689 10.7374 7.50142 10.3714 7.49998 10.0001ZM17.9725 11.2784C16.8608 13.0834 14.3483 16.1209 9.99998 16.1209C8.61642 16.1287 7.25228 15.7952 6.02831 15.1501L7.70581 13.4726C8.50587 14.0038 9.46506 14.2418 10.4206 14.1462C11.3762 14.0507 12.2693 13.6275 12.9483 12.9484C13.6274 12.2694 14.0506 11.3763 14.1461 10.4208C14.2416 9.46519 14.0036 8.506 13.4725 7.70595L15.3975 5.78095C16.4182 6.60618 17.2893 7.60099 17.9725 8.72178C18.2099 9.10591 18.3356 9.54855 18.3356 10.0001C18.3356 10.4517 18.2099 10.8943 17.9725 11.2784Z"
+                fill="#ACACAC"
+              />
+            </svg>
+          ) : (
+            <svg
+              className="passwordToogler passwordToogler_error"
+              width="22"
+              height="16"
+              viewBox="0 0 22 16"
+              fill="none"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                d="M20.3925 5.84907C19.1 3.74407 16.16 0.212402 11 0.212402C5.83998 0.212402 2.89998 3.74407 1.60748 5.84907C1.20792 6.49533 0.996277 7.24011 0.996277 7.99991C0.996277 8.75971 1.20792 9.50449 1.60748 10.1507C2.89998 12.2557 5.83998 15.7874 11 15.7874C16.16 15.7874 19.1 12.2557 20.3925 10.1507C20.792 9.50449 21.0037 8.75971 21.0037 7.99991C21.0037 7.24011 20.792 6.49533 20.3925 5.84907V5.84907ZM18.9716 9.27824C17.8616 11.0832 15.3491 14.1207 11 14.1207C6.65081 14.1207 4.13831 11.0832 3.02831 9.27824C2.79093 8.89411 2.66519 8.45147 2.66519 7.99991C2.66519 7.54834 2.79093 7.1057 3.02831 6.72157C4.13831 4.91657 6.65081 1.87907 11 1.87907C15.3491 1.87907 17.8616 4.91324 18.9716 6.72157C19.209 7.1057 19.3348 7.54834 19.3348 7.99991C19.3348 8.45147 19.209 8.89411 18.9716 9.27824V9.27824Z"
+                fill="#ACACAC"
+              />
+              <path
+                d="M11 3.8335C10.1759 3.8335 9.37031 4.07787 8.68511 4.53571C7.9999 4.99355 7.46585 5.64429 7.15048 6.40565C6.83512 7.16701 6.7526 8.00479 6.91338 8.81304C7.07415 9.62129 7.47098 10.3637 8.0537 10.9464C8.63642 11.5292 9.37885 11.926 10.1871 12.0868C10.9954 12.2475 11.8331 12.165 12.5945 11.8497C13.3559 11.5343 14.0066 11.0002 14.4644 10.315C14.9223 9.62983 15.1666 8.82425 15.1666 8.00016C15.1653 6.8955 14.7259 5.83646 13.9448 5.05535C13.1637 4.27423 12.1046 3.83482 11 3.8335V3.8335ZM11 10.5002C10.5055 10.5002 10.0222 10.3535 9.61106 10.0788C9.19993 9.80413 8.8795 9.41369 8.69028 8.95687C8.50106 8.50006 8.45155 7.99739 8.54802 7.51244C8.64448 7.02748 8.88258 6.58203 9.23221 6.2324C9.58185 5.88276 10.0273 5.64466 10.5123 5.5482C10.9972 5.45174 11.4999 5.50124 11.9567 5.69046C12.4135 5.87968 12.804 6.20011 13.0787 6.61124C13.3534 7.02236 13.5 7.50571 13.5 8.00016C13.5 8.6632 13.2366 9.29909 12.7677 9.76793C12.2989 10.2368 11.663 10.5002 11 10.5002Z"
+                fill="#ACACAC"
+              />
+            </svg>
+          )
+          :
+          null
+        }
+      </div>
+    </div>
+  );
+};
