--- frontend/src/components/signPages/regPage/mailInput.components.tsx
@@ -1,29 +1,0 @@
-import React from "react";
-
-export const MailInput = () => {
-  return (
-    <div id='containerForm' className="form__container ">
-      <label htmlFor="emailInput">
-        <svg
-          id="icon"
-          width="20"
-          height="20"
-          viewBox="0 0 20 20"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            className="form__icon"
-            d="M15.8333 0.833332H4.16666C3.062 0.834655 2.00296 1.27407 1.22184 2.05518C0.440727 2.8363 0.00131558 3.89534 -7.62939e-06 5L-7.62939e-06 15C0.00131558 16.1047 0.440727 17.1637 1.22184 17.9448C2.00296 18.7259 3.062 19.1654 4.16666 19.1667H15.8333C16.938 19.1654 17.997 18.7259 18.7781 17.9448C19.5593 17.1637 19.9987 16.1047 20 15V5C19.9987 3.89534 19.5593 2.8363 18.7781 2.05518C17.997 1.27407 16.938 0.834655 15.8333 0.833332ZM4.16666 2.5H15.8333C16.3323 2.50098 16.8196 2.65126 17.2324 2.9315C17.6453 3.21174 17.9649 3.60912 18.15 4.0725L11.7683 10.455C11.2987 10.9228 10.6628 11.1854 9.99999 11.1854C9.33714 11.1854 8.7013 10.9228 8.23166 10.455L1.84999 4.0725C2.03512 3.60912 2.35467 3.21174 2.76753 2.9315C3.18039 2.65126 3.66767 2.50098 4.16666 2.5ZM15.8333 17.5H4.16666C3.50362 17.5 2.86773 17.2366 2.39889 16.7678C1.93005 16.2989 1.66666 15.663 1.66666 15V6.25L7.05332 11.6333C7.83551 12.4135 8.8952 12.8517 9.99999 12.8517C11.1048 12.8517 12.1645 12.4135 12.9467 11.6333L18.3333 6.25V15C18.3333 15.663 18.0699 16.2989 17.6011 16.7678C17.1322 17.2366 16.4964 17.5 15.8333 17.5Z"
-            fill="#ACACAC"
-          />
-        </svg>
-      </label>
-      <input
-        id="emailInput"
-        className="form__input form__mail form__mail_error"
-        placeholder="Email"
-      />
-    </div>
-  );
-};
