--- frontend/src/components/signPages/regPage/socialBtns.components.tsx
@@ -1,51 +1,0 @@
-import React from "react";
-import "../loginPage/socialBtns.components.scss";
-
-export const SocialBtns = () => {
-  return (
-    <div className="form__socialBtns">
-      <a href="#" className="circle">
-        <svg
-          width="12"
-          height="23"
-          viewBox="0 0 12 23"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M0.903695 11.9302H3.3591V22.0387C3.3591 22.2383 3.52081 22.4 3.72039 22.4H7.88361C8.08318 22.4 8.2449 22.2383 8.2449 22.0387V11.9779H11.0676C11.2511 11.9779 11.4055 11.8401 11.4265 11.6578L11.8552 7.93639C11.867 7.834 11.8345 7.73147 11.766 7.65466C11.6975 7.57778 11.5993 7.53377 11.4964 7.53377H8.24504V5.20099C8.24504 4.49778 8.62368 4.14118 9.37053 4.14118C9.47697 4.14118 11.4964 4.14118 11.4964 4.14118C11.6959 4.14118 11.8577 3.9794 11.8577 3.77989V0.363964C11.8577 0.164387 11.6959 0.00267355 11.4964 0.00267355H8.56666C8.546 0.00166194 8.50011 0 8.43248 0C7.92414 0 6.15722 0.0997883 4.76148 1.38381C3.21501 2.80672 3.42998 4.51042 3.48136 4.80581V7.5337H0.903695C0.704118 7.5337 0.542404 7.69541 0.542404 7.89499V11.5689C0.542404 11.7685 0.704118 11.9302 0.903695 11.9302Z"
-            fill="white"
-          />
-        </svg>
-      </a>
-      <a href="#" className="circle">
-        <svg
-          width="18"
-          height="18"
-          viewBox="0 0 18 18"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M4.31333 7.39001C4.13604 7.90824 4.04593 8.45229 4.04667 9.00001C4.04667 9.61167 4.15667 10.1975 4.36 10.7383C4.61343 11.4143 5.0115 12.0267 5.52636 12.5328C6.04123 13.0388 6.66046 13.4263 7.34071 13.668C8.02097 13.9097 8.74582 13.9998 9.46456 13.9321C10.1833 13.8643 10.8785 13.6404 11.5017 13.2758H11.5025C12.4842 12.7003 13.2385 11.8051 13.6392 10.74H9.18334V7.44334H17.1875C17.3945 8.53374 17.3821 9.65448 17.1508 10.74C16.7502 12.6166 15.7141 14.2973 14.2175 15.4983C12.7393 16.6886 10.8978 17.3363 9 17.3333C7.56914 17.3341 6.16219 16.9664 4.9147 16.2656C3.6672 15.5648 2.62122 14.5546 1.87751 13.3322C1.1338 12.1098 0.717445 10.7164 0.668548 9.28639C0.619651 7.85637 0.939863 6.43785 1.59833 5.16751C2.30081 3.81077 3.36282 2.67331 4.66823 1.87951C5.97364 1.08571 7.47219 0.666136 9 0.666672C11.0217 0.666672 12.8758 1.38667 14.3192 2.58501L11.6192 4.79501C10.9867 4.40137 10.2736 4.15537 9.53294 4.07529C8.79228 3.99521 8.0431 4.08311 7.34109 4.33246C6.63908 4.58181 6.00231 4.98619 5.47812 5.51554C4.95393 6.04489 4.5558 6.68559 4.31333 7.39001Z"
-            fill="white"
-          />
-        </svg>
-      </a>
-      <a href="#" className="circle">
-        <svg
-          width="20"
-          height="12"
-          viewBox="0 0 20 12"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M19.5467 0.819996C19.3217 1.5275 19.0183 2.14333 18.6383 2.7025L18.6542 2.6775C18.2808 3.29972 17.835 4.01861 17.3167 4.83416C16.8722 5.47805 16.6369 5.81138 16.6108 5.83416C16.4958 5.97833 16.4158 6.155 16.3883 6.34916L16.3875 6.355C16.41 6.53166 16.4917 6.68583 16.6108 6.80083L16.9442 7.1725C18.7275 9.00527 19.7306 10.2686 19.9533 10.9625C19.9825 11.0392 19.9992 11.1275 19.9992 11.22C19.9992 11.3817 19.9475 11.5308 19.86 11.6533L19.8617 11.6508C19.715 11.8017 19.51 11.895 19.2833 11.895C19.2583 11.895 19.2325 11.8942 19.2083 11.8917H19.2117H17.0192C17.0183 11.8917 17.0167 11.8917 17.015 11.8917C16.7342 11.8917 16.4758 11.7933 16.2733 11.63L16.2758 11.6317C15.9808 11.3892 15.7183 11.1267 15.4833 10.8417L15.4758 10.8317C15.0919 10.3978 14.7514 10.0325 14.4542 9.73583C13.4642 8.79472 12.7336 8.32416 12.2625 8.32416C12.245 8.3225 12.225 8.32166 12.2042 8.32166C12.0667 8.32166 11.9383 8.36416 11.8325 8.43666L11.835 8.435C11.7483 8.54333 11.6958 8.68333 11.6958 8.83416C11.6958 8.86416 11.6975 8.8925 11.7017 8.92166V8.91833C11.6792 9.18 11.6658 9.48416 11.6658 9.79166C11.6658 9.89083 11.6675 9.98916 11.67 10.0875V10.0733V11.0392C11.6783 11.0783 11.6833 11.1233 11.6833 11.1692C11.6833 11.3708 11.5917 11.5508 11.4483 11.67L11.4475 11.6708C11.125 11.8183 10.7483 11.905 10.3508 11.905C10.2658 11.905 10.1825 11.9008 10.1 11.8933L10.1108 11.8942C8.79917 11.8692 7.5825 11.485 6.54834 10.8375L6.5775 10.8542C5.31834 10.0892 4.275 9.09333 3.48 7.91666L3.4575 7.88166C2.7025 6.88 1.99834 5.74833 1.39834 4.555L1.34 4.42666C0.989169 3.75833 0.625836 2.93666 0.312502 2.09166L0.262502 1.9375C0.135002 1.5575 0.0425024 1.11583 0.00416911 0.65833L0.00250244 0.637497C0.00250244 0.216385 0.25028 0.00582997 0.745836 0.00582997H2.9375C2.9575 0.0041633 2.98084 0.00332975 3.005 0.00332975C3.21167 0.00332975 3.4025 0.0741632 3.55334 0.19333L3.55167 0.191663C3.71834 0.374163 3.84167 0.59833 3.9025 0.84833L3.905 0.85833C4.30834 1.9975 4.73584 2.95416 5.2275 3.87083L5.17167 3.75666C5.57334 4.565 6.0175 5.26083 6.52417 5.90333L6.505 5.8775C6.94945 6.42305 7.29611 6.69583 7.545 6.69583C7.5525 6.69666 7.56167 6.69666 7.57167 6.69666C7.695 6.69666 7.8025 6.63 7.86084 6.53L7.86167 6.52833C7.92334 6.37 7.95917 6.18583 7.95917 5.99416C7.95917 5.95333 7.9575 5.9125 7.95417 5.8725V5.8775V2.645C7.93667 2.26416 7.84334 1.90833 7.6875 1.58916L7.69417 1.605C7.59084 1.385 7.46667 1.195 7.31917 1.02583L7.32167 1.02916C7.185 0.892496 7.09167 0.71333 7.0625 0.51333L7.06167 0.50833C7.06167 0.366663 7.12667 0.239163 7.22834 0.15583L7.22917 0.154997C7.33 0.0624965 7.465 0.00666324 7.6125 0.00666324H7.61917H11.0742C11.0925 0.00416324 11.1133 0.00332975 11.135 0.00332975C11.2975 0.00332975 11.4433 0.0766633 11.54 0.191663L11.5408 0.192497C11.6267 0.349163 11.6775 0.536663 11.6775 0.734997C11.6775 0.766663 11.6758 0.797497 11.6733 0.82833V0.824163V5.135C11.6717 5.155 11.6708 5.17833 11.6708 5.20166C11.6708 5.355 11.7133 5.49916 11.7867 5.62166L11.785 5.61833C11.8458 5.70833 11.9475 5.76666 12.0625 5.76666C12.2092 5.75666 12.3433 5.70833 12.4558 5.63166L12.4533 5.63333C12.68 5.4775 12.875 5.3 13.0442 5.09916L13.0475 5.095C13.5533 4.52333 14.02 3.89166 14.4267 3.22L14.4592 3.16166C14.7458 2.68666 15.0558 2.10166 15.3367 1.50083L15.3883 1.37833L15.76 0.634996C15.8892 0.26333 16.2358 0.00166321 16.6433 0.00166321C16.6592 0.00166321 16.675 0.00166315 16.6908 0.00249648H16.6883H18.8808C19.4731 0.00249648 19.6944 0.274996 19.5467 0.819996Z"
-            fill="white"
-          />
-        </svg>
-      </a>
-    </div>
-  );
-};
