+++ frontend/src/components/signPages/socialBtns.components.scss
@@ -1,4 +1,4 @@
-@import "../../../config.scss";
+@import "../../config.scss";
 
 .form__socialBtns {
   margin-top: 17px;
