+++ frontend/src/pages/signPages/login.pages.tsx
@@ -1,10 +1,10 @@
 import { Navigate } from 'react-router-dom';
-import { Form } from '../../components/signPages/loginPage/form.components';
+import { SignForm } from '../../components/signPages/form.components';
 
 export const LoginPage = (): JSX.Element | null => {
 
   if (localStorage.getItem('token') === null) {
-    return <Form />
+    return <SignForm page={'logIn'} title={'Welcome to EngLearn'} buttonContent={'Log in'} />
   } else {
     return <Navigate to={'/'}></Navigate>
   }
