+++ frontend/src/utils/export.utils.ts
@@ -1,0 +1,2 @@
+export { LoginChecker } from './signPages/loginPage/loginChecker.utils';
+export { RegChecker } from './signPages/regPage/regChecker.utils';
